Change file location, comment unnecessary code
