+++ Bomba-Challenge1/Model/GameData.swift
@@ -7,66 +7,66 @@
 
 import Foundation
 
-class GameData {
-    
-    static let shared = GameData()
-    
-    var categories: [Category]
-    let punishments: [String]
-    lazy var questionArray: [String] = {
-        var array = [String]()
-        self.categories.forEach { category in
-            if category.isSelected {
-                array += category.questions
-            }
-        }
-        return array
-    }()
-    
-    private init() {
-        let generalCategory = Category(name: "О разном",
-                                       imageName: "category1",
-                                       questions: ["Вопрос1",
-                                                   "Вопрос2"
-                                                  ])
-        let sportCategory = Category(name: "Спорт и Хобби",
-                                     imageName: "category2",
-                                     questions: ["Вопрос1",
-                                                 "Вопрос2"
-                                                ])
-        let lifeCategory = Category(name: "Про Жизнь",
-                                    imageName: "category3",
-                                    questions: ["Вопрос1",
-                                                "Вопрос2"
-                                               ])
-        let celebrityCategory = Category(name: "Знаменитости",
-                                         imageName: "category4",
-                                         questions: ["Вопрос1",
-                                                     "Вопрос2"
-                                                    ])
-        let movieCategory = Category(name: "Искусство и Кино",
-                                     imageName: "category5",
-                                     questions: ["Вопрос1",
-                                                 "Вопрос2"
-                                                ])
-        let natureCategory = Category(name: "Природа",
-                                      imageName: "category6",
-                                      questions: ["Вопрос1",
-                                                  "Вопрос2"
-                                                 ])
-        categories = [generalCategory, sportCategory, lifeCategory, celebrityCategory, movieCategory, natureCategory]
-        punishments = [
-            "Наказание 1",
-            "Наказание2",
-            "Наказание3",
-            "Наказание4",
-            "Наказание5",
-            "Наказание6",
-            "Наказание7",
-            "Наказание8",
-            "Наказание9",
-            "Наказание10",
-        ]
-    }
-    
-}
+//class GameData {
+//    
+//    static let shared = GameData()
+//    
+//    var categories: [Category]
+//    let punishments: [String]
+//    lazy var questionArray: [String] = {
+//        var array = [String]()
+//        self.categories.forEach { category in
+//            if category.isSelected {
+//                array += category.questions
+//            }
+//        }
+//        return array
+//    }()
+//    
+//    private init() {
+//        let generalCategory = Category(name: "О разном",
+//                                       imageName: "category1",
+//                                       questions: ["Вопрос1",
+//                                                   "Вопрос2"
+//                                                  ])
+//        let sportCategory = Category(name: "Спорт и Хобби",
+//                                     imageName: "category2",
+//                                     questions: ["Вопрос1",
+//                                                 "Вопрос2"
+//                                                ])
+//        let lifeCategory = Category(name: "Про Жизнь",
+//                                    imageName: "category3",
+//                                    questions: ["Вопрос1",
+//                                                "Вопрос2"
+//                                               ])
+//        let celebrityCategory = Category(name: "Знаменитости",
+//                                         imageName: "category4",
+//                                         questions: ["Вопрос1",
+//                                                     "Вопрос2"
+//                                                    ])
+//        let movieCategory = Category(name: "Искусство и Кино",
+//                                     imageName: "category5",
+//                                     questions: ["Вопрос1",
+//                                                 "Вопрос2"
+//                                                ])
+//        let natureCategory = Category(name: "Природа",
+//                                      imageName: "category6",
+//                                      questions: ["Вопрос1",
+//                                                  "Вопрос2"
+//                                                 ])
+//        categories = [generalCategory, sportCategory, lifeCategory, celebrityCategory, movieCategory, natureCategory]
+//        punishments = [
+//            "Наказание 1",
+//            "Наказание2",
+//            "Наказание3",
+//            "Наказание4",
+//            "Наказание5",
+//            "Наказание6",
+//            "Наказание7",
+//            "Наказание8",
+//            "Наказание9",
+//            "Наказание10",
+//        ]
+//    }
+//    
+//}
