+++ Bomba-Challenge1/Model/Punishments.swift
